requirements.txt
django-markdownify - 0.8.0